mlflow/src/config.py
import os
from dotenv import load_dotenv

load_dotenv()

MLFLOW_S3_ENDPOINT_URL = os.getenv('MLFLOW_S3_ENDPOINT_URL')
AWS_ACCESS_KEY_ID = os.getenv('AWS_ACCESS_KEY_ID')
AWS_SECRET_ACCESS_KEY = os.getenv('AWS_SECRET_ACCESS_KEY')

POSTGRES_USER = os.getenv('POSTGRES_USER')
POSTGRES_PASSWORD = os.getenv('POSTGRES_PASSWORD')
POSTGRES_DB = os.getenv('POSTGRES_DB')

MLFLOW_TRACKING_URI = "http://localhost:5000"  
MLFLOW_EXPERIMENT_NAME = "Car_Price_Prediction"
